Adds some optimizations
