+++ README.md
@@ -1,3 +1,8 @@
 # Rick and Morty API
 
 This repo will be used to help us learn how to develop our own APIs using Ruby on Rails. The goal at the end of this will be to recreate the Rick and Morty api from https://rickandmortyapi.com/documentation/#location-schema. 
+
+## Run
+In order to run the project you can launch the following command:
+`./bootstrap.sh`
+which starts rails and the racecar consumers in one shot
