+++ bootstrap.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+rails s &
+bundle exec racecar RawDataConsumer &
+bundle exec racecar TransformerConsumer
