+++ lib/avro/schemas.avsc
